+++ govtool/frontend/src/hooks/forms/index.ts
@@ -1,5 +1,4 @@
 export * from "./useCreateGovernanceActionForm";
-export * from "./useDelegateTodRepForm";
 export * from "./useEditDRepInfoForm";
 export * from "./useRegisterAsdRepForm";
 export * from "./useVoteActionForm";