+++ govtool/frontend/src/services/requests/getAdaHolderCurrentDelegation.ts
@@ -1,6 +1,7 @@
 import { CurrentDelegation } from "@models";
 
 import { API } from "../API";
+import { fixViewForScriptBasedDRep } from "@/utils";
 
 export const getAdaHolderCurrentDelegation = async ({
   stakeKey,
@@ -11,5 +12,16 @@ export const getAdaHolderCurrentDelegation = async ({
     `/ada-holder/get-current-delegation/${stakeKey}`,
   );
 
-  return response.data;
+  if (!response.data) return response.data;
+
+  // DBSync contains wrong representation of DRep view for script based DReps
+  const view = response.data.dRepView && fixViewForScriptBasedDRep(
+    response.data.dRepView,
+    response.data.isDRepScriptBased,
+  );
+
+  return {
+    ...response.data,
+    dRepView: view,
+  };
 };